requirements.txt
streamlit==1.23.1
tensorflow==2.15.0
opencv-python-headless==4.8.0.74
pillow==9.6.1
numpy==1.24.2
